rename `buckets` module into `bucket_management`
