+++ src/sts/bucket_management/minio.py
@@ -3,7 +3,7 @@ from itertools import chain
 from minio import S3Error
 
 import sts.logs
-from sts.buckets.service import BucketService
+from sts.bucket_management.service import BucketService
 from sts.config import AppSettings
 from sts.file_storage.client import FileStorageClient
 from sts.models.bucket import BucketsInfo
